WIP: Attempting to make the HOST Flask webapp visible to the minion_worker container (the one that processes NextML requests) so that it may add and remove from the Flask MongoDb database.
net: "bridge" is supposed to bridge the Docker network with
the host network. The fix might be as simple as putting
the host network onto/into the Docker network but research
woudl have to be done to resolve that.

This would allow the Metadata Relevance oracle to remove
submitted business information that has been verified to be true
and allow incorrect business informatoin to remain in the pool
of businesses for Mturkers to try again on.